Tweaked update method so if ports is specified as an empty array, existing rules are not deleted.
#595-patch-ecloud-v2-firewall-rules-firewallruleid-issues

+++ app/Http/Controllers/V2/FirewallRuleController.php
@@ -110,7 +110,8 @@ class FirewallRuleController extends BaseController
         ]));
         $firewallRule->save();
 
-        if ($request->has('ports')) {
+        if ($request->has('ports') &&
+            (is_array($request->get('ports') && count($request->get('ports') > 0)))) {
             $firewallRule->firewallRulePorts->each(function ($rule) {
                 $rule->delete();
             });
